Move load_pandas() to utils
